Add test for fixture exception
