+++ Src/AutoFixture.NUnit3.UnitTest/AutoFixture.NUnit3.UnitTest.csproj
@@ -43,6 +43,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="InlineAutoDataAttributeStub.cs" />
     <Compile Include="AutoDataAttributeStub.cs" />
     <Compile Include="AutoDataAttributeTest.cs" />
     <Compile Include="CustomizeAttributeTest.cs" />
@@ -58,6 +59,7 @@
     <Compile Include="MyClass.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Scenario.cs" />
+    <Compile Include="ThrowingStubFixture.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\AutoFixture.NUnit3\AutoFixture.NUnit3.csproj">
